Support IPsbType as plugin
