+++ FreeMote.Psb/FreeMote.Psb.csproj
@@ -54,16 +54,21 @@
     <Compile Include="Plugins\IPsbKeyProvider.cs" />
     <Compile Include="Plugins\IPsbPluginInfo.cs" />
     <Compile Include="Plugins\IPsbShell.cs" />
+    <Compile Include="Plugins\IPsbSpecialType.cs" />
     <Compile Include="PSB.cs" />
     <Compile Include="PsbHelper.cs" />
     <Compile Include="PsbPainter.cs" />
     <Compile Include="PsbResCollector.cs" />
-    <Compile Include="PsbTypes.cs" />
+    <Compile Include="PsbValues.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ImageMetadata.cs" />
+    <Compile Include="Types\ArchiveType.cs" />
+    <Compile Include="Types\FontType.cs" />
     <Compile Include="Types\ImageType.cs" />
     <Compile Include="Types\MmoType.cs" />
     <Compile Include="Types\MotionType.cs" />
+    <Compile Include="Types\PimgType.cs" />
+    <Compile Include="Types\ScnType.cs" />
     <Compile Include="Types\SoundArchiveType.cs" />
     <Compile Include="Textures\TextureCombiner.cs" />
     <Compile Include="Textures\TexturePacker.cs" />
