+++ FreeMote.Psb/Plugins/IPsbSpecialType.cs
@@ -1,0 +1,9 @@
+﻿using FreeMote.Psb;
+
+namespace FreeMote.Plugins
+{
+    public interface IPsbSpecialType : IPsbType, IPsbPlugin
+    {
+        string TypeId { get; }
+    }
+}
