+++ FreeMote.Psb/Types/ArchiveType.cs
@@ -1,0 +1,11 @@
+﻿namespace FreeMote.Psb.Types
+{
+    class ArchiveType : IPsbType
+    {
+        public PsbType PsbType => PsbType.ArchiveInfo;
+        public bool IsThisType(PSB psb)
+        {
+            return psb.TypeId == "archive"; //&& psb.Objects.ContainsKey("file_info")
+        }
+    }
+}
