+++ FreeMote/Consts.cs
@@ -100,6 +100,11 @@ namespace FreeMote
         /// Allows you to edit CI* images by re-generate the palette for each bppIndexed image (will increase size), otherwise you should not change those images
         /// </summary>
         public static bool GeneratePalette { get; set; } = true;
+
+        /// <summary>
+        /// If true, prefer plugins than inherit
+        /// </summary>
+        public static bool PluginFirst { get; set; } = false;
     }
 
     //REF: https://stackoverflow.com/a/24987840/4374462
